[WIP] generalize the API
