+++ exts/loupe.simulation.beckhoff_bridge/loupe/simulation/common/Global.py
@@ -1,0 +1,2 @@
+
+LOUPE_MENU_ITEM_NAME = "Loupe"
